requirements.txt
django-autoslug==1.9.8
django-crispy-forms
django-icons
Pillow